Drop python 2.6 support and add python 3 support
Dropped unittest2 since only the 2.6 tests required them, and it
makes supporting Python 2 and 3 in a single codebase simpler.

Added `six` requirement so `iteritems` and `itervalues` is still
used on Python 2.

Added `tox` which makes testing multiple python versions locally a lot
easier.

Moved `nosetests` settings to `setup.cfg`, so it doesn't have to be
specified in multiple places, and add it's collector as the `test_suite`
so people can use `python setup.py test` and `python setup.py nosetests`