Make the nested results behavior the default
Demands is for wrapping service endpoints, a paginated endpoint should
always return a dict, not a list, so that pagination meta information
can be provided. This is how all of our services work at yola. So this
commit makes that the default behavior, while allowing the "results" key
to be customized, or set to None for the original behavior of a
top-level list.

This is a backwards-incompatible change. Consumers already using a
pagination iterator will need to pass results_key=None to get the old
behavior, or update their code to handle the new behavior (ideally that
would mean removing whatever wrapper function they were using to handle
the paginated service calls).